Account for Rack's broken use of string encodings (#5)
Rack assigns 8-bit ASCII encoding to all strings in its request environment. This causes issues when these strings have characters with the 8th bit set and are converted to UTF-8 for output to the HAR file. Work around this by forcing all strings taken from the request environment to be UTF-8.

https://github.com/phusion/passenger/issues/1328

This bumps the version number to 0.2.3.