Ignore binary data in response bodies (#9)
If a response body has binary data, ignore the body but record its length. A response body is considered to have binary data if it has an invalid source encoding (other than 8-bit ASCII) or if it cannot be represented in UTF-8.

Fix: Don't assume `fixEncoding` is given a string. If it is given a non-string, log a warning and return the argument unchanged.

Assume UTF-8 by default for request bodies.

Bump to v0.2.6.