Work around non-Rack-compliant middleware
Some middleware is not Rack-compliant and closes the request input
stream, even though the spec says not to. Work around this by copying
the contents of the request input stream before calling down the
middleware stack, and then swapping the copy into the Rack environment
before calling into the logger.